增加断点续传cancel时是否清除上传事件的设置项
Conflicts:
	oss-android-sdk/src/androidTest/java/com/alibaba/sdk/android/OSSTestConfig.java

+++ oss-android-sdk/src/main/java/com/alibaba/sdk/android/oss/internal/ExtensionRequestOperation.java
@@ -102,7 +102,6 @@ public class ExtensionRequestOperation {
         public ResumableUploadResult call() throws Exception {
 
             try {
-
                 initUploadId();
                 ResumableUploadResult result = doMultipartUpload();
 
@@ -185,9 +184,11 @@ public class ExtensionRequestOperation {
         private ResumableUploadResult doMultipartUpload() throws IOException, ClientException, ServiceException {
 
             if (context.getCancellationHandler().isCancelled()) {
-                abortThisResumableUpload();
-                if (recordFile != null) {
-                    recordFile.delete();
+                if (request.deleteUploadOnCancelling()) {
+                    abortThisResumableUpload();
+                    if (recordFile != null) {
+                        recordFile.delete();
+                    }
                 }
                 throwOutInterruptClientException();
             }
@@ -243,9 +244,11 @@ public class ExtensionRequestOperation {
                 currentUploadIndex++;
 
                 if (context.getCancellationHandler().isCancelled()) {
-                    abortThisResumableUpload();
-                    if (recordFile != null) {
-                        recordFile.delete();
+                    if (request.deleteUploadOnCancelling()) {
+                        abortThisResumableUpload();
+                        if (recordFile != null) {
+                            recordFile.delete();
+                        }
                     }
                     throwOutInterruptClientException();
                 }
